refactor challenge v3
